+++ core/core/build.h
@@ -7,14 +7,14 @@
 
 struct buildVector
 {
-	short lowLimit;
-	short hiLimit;
-	unsigned short vecOffset;
+	size_t lowLimit;
+	size_t hiLimit;
+	size_t vecOffset;
 	std::vector<token::baseToken*>* vecPtr;
-	buildVector(short lowLimit, short hiLimit, short vecOffset, std::vector<token::baseToken*>* vecPtr);
+	buildVector(size_t lowLimit, size_t hiLimit, size_t vecOffset, std::vector<token::baseToken*>* vecPtr);
 	buildVector();
 	//Return types, -1 error, >0 next operator index
-	short calculateNextOperation(short lowLimit, short hiLimit);
+	size_t calculateNextOperation(size_t lowLimit, size_t hiLimit);
 
 };
